\name{qrandom}
\alias{qrandom}
\title{...}
\description{
  \code{qrandom}

}
\usage{
qrandom(n=1, type="uint8", blocksize=1)
}
\arguments{
  \item{n}{an object of class \code{"\link{}"} (or one that
    can be coerced to that class)}

  \item{type}{}

  \item{blocksize}{}
}

\value{
  \code{qrandom} returns an object of \code{\link{class}} \code{..}.

  The returning value of \code{"qrandom"} is a \code{"..."} containing the
  following components:

  \item{...}{...}

}


\references{

}

\examples{

}
