\name{qrandomunif}
\alias{qrandomunif}
\title{...}
\description{
  \code{qrandomunif}

}
\usage{
qrandomunif(n=1, type="uint8", blocksize=1)
}
\arguments{
  \item{n}{an object of class \code{"\link{}"} (or one that
    can be coerced to that class)}

  \item{type}{}

  \item{blocksize}{}
}

\value{
  \code{qrandomunif} returns an object of \code{\link{class}} \code{..}.

  The returning value of \code{"qrandomunif"} is a \code{"..."} containing the
  following components:

  \item{...}{...}

}


\references{

}

\examples{

}
